Reorganize folder structure Modify screens for some tweaks
This commit includes a major refactoring of the codebase. The screens related files have been moved to a separate 'screens' directory for better organization. The import statements in main.py have been updated to reflect this change.
